attempt 4

+++ src/main/java/unit2/Unit2.java
@@ -15,7 +15,7 @@ public class Unit2 {
      */
     public static String concatenateStrings(String str1, String str2) {
         // TODO: Implement this method using String concatenation
-        return str1 + " " + str2;
+        return str1+" "+str2;
     }
 
     /**
@@ -68,7 +68,7 @@ public class Unit2 {
      */
     public static int roundValue(double value) {
         // TODO: Implement this method using Math.round
-        return Math.round(value);
+        return (int)Math.round(value);
     }
 
     /**
